refactor: moved categories types to external file
